Reorganize logging code a bit
