+++ src/Utils/Logging/Config.hs
@@ -17,7 +17,7 @@
 -- Maintainer: Lars Kuhtz <lars@kadena.io>
 -- Stability: experimental
 --
--- TODO
+-- Configuration for generic logging utils
 --
 module Utils.Logging.Config
 (
@@ -43,21 +43,6 @@ module Utils.Logging.Config
 , pBackendConfig
 , pBackendConfig_
 
--- * Amberdata Logger Backend
-, AmberdataConfig(..)
-  
--- * Logging Config
-, ClusterId
-, LogConfig(..)
-, logConfigLogger
-, logConfigBackend
-, logConfigTelemetryBackend
-, logConfigClusterId
-, logConfigAmberdataBackend
-, defaultLogConfig
-, validateLogConfig
-, pLogConfig
-, pLogConfig_
 ) where
 
 import Configuration.Utils
@@ -67,7 +52,6 @@ import Control.DeepSeq
 import Control.Lens.TH
 import Control.Monad.Catch
 import Control.Monad.Writer
-import Data.Bool (bool)
 
 import qualified Data.CaseInsensitive as CI
 import Data.String
@@ -76,7 +60,6 @@ import qualified Data.Text as T
 import GHC.Generics
 
 import System.Logger.Backend.ColorOption
-import System.Logger.Logger
 
 -- internal modules
 
@@ -245,143 +228,3 @@ pBackendConfig_ prefix = id
     <*< backendConfigFormat .:: pLogFormat_ prefix
     <*< backendConfigHandle .:: pHandleConfig_ prefix
 
--- -------------------------------------------------------------------------- --
--- Logging System Configuration
-
--- | An user provided label that is used to tag log messages from nodes.
---
--- It intended purpose is to allow users to defines arbitrary groups of nodes
--- and tag the respective log messages.
---
-newtype ClusterId = ClusterId T.Text
-    deriving (Show, Eq, Ord, Generic)
-    deriving newtype (IsString, ToJSON, FromJSON)
-    deriving anyclass (NFData)
-
-instance HasTextRepresentation ClusterId where
-    toText (ClusterId t) = t
-    fromText = return . ClusterId
-
-
--- | Backend for logging to Amberdata
---
-data AmberdataConfig
-  = AmberdataConfig
-    { _ambredataConfigHost :: HostAddress
-    , _amberdataApiKey :: T.Text
-    , _amberdataBlockchainId :: T.Text
-    , _amberdataDebug :: Bool
-    }
-  deriving (Show, Eq, Ord, Generic)
-
-amberdataConfigFromText :: MonadThrow m => T.Text -> m AmberdataConfig
-amberdataConfigFromText x
-  | CI.mk (T.take (T.length hostPrefix) x) == CI.mk hostPrefix =
-      case T.splitOn "::" x of
-        [hostStr, api, bid, "debug"] -> formatConfig hostStr api bid True
-        [hostStr, api, bid] -> formatConfig hostStr api bid False
-        _ -> configFromTextErr $ CI.mk x
-  | otherwise = configFromTextErr $ CI.mk x
-
-  where hostPrefix = "amberdata:"
-
-        configFromTextErr e =
-          throwM $ DecodeException $ "unexpected logger handle value: "
-          <> fromString (show e)
-          <> ", expected \"amberdata:<HOST>:<PORT>::<APIKEY>::<BLOCKCHAINID>\""
-          <> " or \"amberdata:<HOST>:<PORT>::<APIKEY>::<BLOCKCHAINID>::debug\""
-
-        formatConfig hostStr api bid doDebug = do
-          hostAddr <- fromText $ T.drop (T.length hostPrefix) hostStr
-          return $ AmberdataConfig hostAddr api bid doDebug
-
-amberdataConfigToText :: AmberdataConfig -> T.Text
-amberdataConfigToText (AmberdataConfig serv api bid debugResp)
-  = "amberdata:" <> toText serv
-    <> "::" <> api
-    <> "::" <> bid
-    <> bool "" ("::" <> "debug") debugResp
-
-instance HasTextRepresentation AmberdataConfig where
-    toText = amberdataConfigToText
-    fromText = amberdataConfigFromText
-
-    {-# INLINE toText #-}
-    {-# INLINE fromText #-}
-
-instance ToJSON AmberdataConfig where
-    toJSON = String . amberdataConfigToText
-
-instance FromJSON AmberdataConfig where
-    parseJSON = parseJsonFromText "AmberdataConfig"
-
-
--- | General logging config
---
-data LogConfig = LogConfig
-    { _logConfigLogger :: !LoggerConfig
-    , _logConfigBackend :: !BackendConfig
-    , _logConfigTelemetryBackend :: !(EnableConfig BackendConfig)
-    , _logConfigAmberdataBackend :: !(Maybe AmberdataConfig)
-    , _logConfigClusterId :: !(Maybe ClusterId)
-    }
-    deriving (Show, Eq, Ord, Generic)
-
-makeLenses ''LogConfig
-
-defaultLogConfig :: LogConfig
-defaultLogConfig = LogConfig
-    { _logConfigLogger = defaultLoggerConfig
-    , _logConfigBackend = defaultBackendConfig
-    , _logConfigTelemetryBackend = defaultEnableConfig defaultBackendConfig
-    , _logConfigAmberdataBackend = Nothing
-      -- ^ Amberdata logging disabled by default
-    , _logConfigClusterId = Nothing
-    }
-
-validateLogConfig :: ConfigValidation LogConfig []
-validateLogConfig o = do
-    validateLoggerConfig $ _logConfigLogger o
-    validateBackendConfig $ _logConfigBackend o
-    validateBackendConfig $ _enableConfigConfig $ _logConfigTelemetryBackend o
-
-instance ToJSON LogConfig where
-    toJSON o = object
-        [ "logger" .= _logConfigLogger o
-        , "backend" .= _logConfigBackend o
-        , "telemetryBackend" .= _logConfigTelemetryBackend o
-        , "clusterId" .= _logConfigClusterId o
-        , "amberdataBackend" .= _logConfigAmberdataBackend o
-        ]
-
-instance FromJSON (LogConfig -> LogConfig) where
-    parseJSON = withObject "LogConfig" $ \o -> id
-        <$< logConfigLogger %.: "logger" % o
-        <*< logConfigBackend %.: "backend" % o
-        <*< logConfigTelemetryBackend %.: "telemetryBackend" % o
-        <*< logConfigClusterId ..: "clusterId" % o
-        <*< logConfigAmberdataBackend ..: "amberdataBackend" % o
-
-pLogConfig :: MParser LogConfig
-pLogConfig = pLogConfig_ ""
-
--- | A version of 'pLogConfig' that takes a prefix for the command
--- line option.
---
-pLogConfig_
-    :: T.Text
-        -- ^ prefix for this and all subordinate command line options.
-    -> MParser LogConfig
-pLogConfig_ prefix = id
-    <$< logConfigLogger %:: pLoggerConfig_ prefix
-    <*< logConfigBackend %:: pBackendConfig_ prefix
-    <*< logConfigTelemetryBackend %::
-        pEnableConfig "telemetry-logger" (pBackendConfig_ $ "telemetry-" <> prefix)
-    <*< logConfigClusterId .:: fmap Just % textOption
-        % long "cluster-id"
-        <> help "a label that is added to all log messages from this node"
-    <*< logConfigAmberdataBackend .:: fmap Just % textOption
-        % long (T.unpack prefix <> "amberdata-config")
-        <> metavar ("amberdata:<HOST>:<PORT>::<APIKEY>::<BLOCKCHAINID>"
-                <> " or amberdata:<HOST>:<PORT>::<APIKEY>::<BLOCKCHAINID>::debug")
-        <> help "config for logging to amberdata"
